Match now works properly tho duplicates still a problem: playlister class will skip any files that have already appeared to get round this. Ran the meta extractor for data and created the cache file in samplemusicxml.

+++ implementation/primaries/Loading/classes/Measure.py
@@ -4,7 +4,7 @@ class Measure(BaseClass.Base):
     def __init__(self, **kwargs):
         BaseClass.Base.__init__(self)
         if "width" in kwargs:
-            self.width = int(kwargs["width"])
+            self.width = float(kwargs["width"])
         self.notes = []
         self.directions = []
